Codex: Cohesion Refactor (#3336)
* Codex: Cohesion Refactor

* refactor: centralize performance fixture path filters

---------
